+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 ## Agile Downloads (Lite)
 
+### Version 1.2.0 
+#### October 2023
+- Improved PHP 8.2 compatibility
+- Improved Joomla 4.x compatibility
+- Added Joomla 5 compatiblity
+- Fixed a Bug where the folder selection modal was not working in certain cases
+
 ### Version 1.0.0
 - Initial Release